+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * 22.0.0
+  - Strongly convex functionality in TotalVariation and FGP_TV Functions.
   - Refactored KullbackLeibler function class. Fix bug on gradient method for SIRF objects.
   - Numba added as a CIL requirement.
   - Simplify initialisation of `CentreOfRotation.ImageSharpness` with new parameter `backend`